feat: add Filter enum and refactor imports for improved filtering capabilities
