+++ examples/table_by_id.rs
@@ -1,6 +1,6 @@
 use std::env;
 
-use baserow_rs::{Baserow, ConfigBuilder, Filter, OrderDirection};
+use baserow_rs::{filter::Filter, Baserow, ConfigBuilder, OrderDirection};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
